+++ model/map/src/main/java/org/keycloak/models/map/client/MapClientEntity.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Red Hat, Inc. and/or its affiliates
+ * Copyright 2021 Red Hat, Inc. and/or its affiliates
  * and other contributors as indicated by the @author tags.
  * 
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -19,460 +19,177 @@ package org.keycloak.models.map.client;
 import org.keycloak.models.ProtocolMapperModel;
 import org.keycloak.models.map.common.AbstractEntity;
 import java.util.Collection;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.Map.Entry;
-import java.util.Objects;
 import java.util.Set;
-import java.util.function.Function;
-import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 /**
  *
  * @author hmlnarik
  */
-public class MapClientEntity<K> implements AbstractEntity<K> {
-
-    private K id;
-    private String realmId;
-
-    private String clientId;
-    private String name;
-    private String description;
-    private Set<String> redirectUris = new HashSet<>();
-    private boolean enabled;
-    private boolean alwaysDisplayInConsole;
-    private String clientAuthenticatorType;
-    private String secret;
-    private String registrationToken;
-    private String protocol;
-    private Map<String, List<String>> attributes = new HashMap<>();
-    private Map<String, String> authFlowBindings = new HashMap<>();
-    private boolean publicClient;
-    private boolean fullScopeAllowed;
-    private boolean frontchannelLogout;
-    private int notBefore;
-    private Set<String> scope = new HashSet<>();
-    private Set<String> webOrigins = new HashSet<>();
-    private Map<String, ProtocolMapperModel> protocolMappers = new HashMap<>();
-    private Map<String, Boolean> clientScopes = new HashMap<>();
-    private Set<String> scopeMappings = new LinkedHashSet<>();
-    private boolean surrogateAuthRequired;
-    private String managementUrl;
-    private String rootUrl;
-    private String baseUrl;
-    private boolean bearerOnly;
-    private boolean consentRequired;
-    private boolean standardFlowEnabled;
-    private boolean implicitFlowEnabled;
-    private boolean directAccessGrantsEnabled;
-    private boolean serviceAccountsEnabled;
-    private int nodeReRegistrationTimeout;
-
-    /**
-     * Flag signalizing that any of the setters has been meaningfully used.
-     */
-    protected boolean updated;
-
-    protected MapClientEntity() {
-        this.id = null;
-        this.realmId = null;
-    }
-
-    public MapClientEntity(K id, String realmId) {
-        Objects.requireNonNull(id, "id");
-        Objects.requireNonNull(realmId, "realmId");
-
-        this.id = id;
-        this.realmId = realmId;
-    }
-
-    @Override
-    public K getId() {
-        return this.id;
-    }
-
-    @Override
-    public boolean isUpdated() {
-        return this.updated;
-    }
-
-    public String getClientId() {
-        return clientId;
-    }
-
-    public void setClientId(String clientId) {
-        this.updated |= ! Objects.equals(this.clientId, clientId);
-        this.clientId = clientId;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.updated |= ! Objects.equals(this.name, name);
-        this.name = name;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.updated |= ! Objects.equals(this.description, description);
-        this.description = description;
-    }
-
-    public Set<String> getRedirectUris() {
-        return redirectUris;
-    }
-
-    public void setRedirectUris(Set<String> redirectUris) {
-        this.updated |= ! Objects.equals(this.redirectUris, redirectUris);
-        this.redirectUris = redirectUris;
-    }
-
-    public boolean isEnabled() {
-        return enabled;
-    }
-
-    public void setEnabled(boolean enabled) {
-        this.updated |= ! Objects.equals(this.enabled, enabled);
-        this.enabled = enabled;
-    }
-
-    public boolean isAlwaysDisplayInConsole() {
-        return alwaysDisplayInConsole;
-    }
-
-    public void setAlwaysDisplayInConsole(boolean alwaysDisplayInConsole) {
-        this.updated |= ! Objects.equals(this.alwaysDisplayInConsole, alwaysDisplayInConsole);
-        this.alwaysDisplayInConsole = alwaysDisplayInConsole;
-    }
-
-    public String getClientAuthenticatorType() {
-        return clientAuthenticatorType;
-    }
-
-    public void setClientAuthenticatorType(String clientAuthenticatorType) {
-        this.updated |= ! Objects.equals(this.clientAuthenticatorType, clientAuthenticatorType);
-        this.clientAuthenticatorType = clientAuthenticatorType;
-    }
-
-    public String getSecret() {
-        return secret;
-    }
-
-    public void setSecret(String secret) {
-        this.updated |= ! Objects.equals(this.secret, secret);
-        this.secret = secret;
-    }
-
-    public String getRegistrationToken() {
-        return registrationToken;
-    }
-
-    public void setRegistrationToken(String registrationToken) {
-        this.updated |= ! Objects.equals(this.registrationToken, registrationToken);
-        this.registrationToken = registrationToken;
-    }
-
-    public String getProtocol() {
-        return protocol;
-    }
-
-    public void setProtocol(String protocol) {
-        this.updated |= ! Objects.equals(this.protocol, protocol);
-        this.protocol = protocol;
-    }
-
-    public Map<String, List<String>> getAttributes() {
-        return attributes;
-    }
-
-    public void setAttribute(String name, List<String> values) {
-        this.updated |= ! Objects.equals(this.attributes.put(name, values), values);
-    }
-
-    public Map<String, String> getAuthFlowBindings() {
-        return authFlowBindings;
-    }
-
-    public void setAuthFlowBindings(Map<String, String> authFlowBindings) {
-        this.updated |= ! Objects.equals(this.authFlowBindings, authFlowBindings);
-        this.authFlowBindings = authFlowBindings;
-    }
-
-    public boolean isPublicClient() {
-        return publicClient;
-    }
-
-    public void setPublicClient(boolean publicClient) {
-        this.updated |= ! Objects.equals(this.publicClient, publicClient);
-        this.publicClient = publicClient;
-    }
-
-    public boolean isFullScopeAllowed() {
-        return fullScopeAllowed;
-    }
-
-    public void setFullScopeAllowed(boolean fullScopeAllowed) {
-        this.updated |= ! Objects.equals(this.fullScopeAllowed, fullScopeAllowed);
-        this.fullScopeAllowed = fullScopeAllowed;
-    }
-
-    public boolean isFrontchannelLogout() {
-        return frontchannelLogout;
-    }
-
-    public void setFrontchannelLogout(boolean frontchannelLogout) {
-        this.updated |= ! Objects.equals(this.frontchannelLogout, frontchannelLogout);
-        this.frontchannelLogout = frontchannelLogout;
-    }
-
-    public int getNotBefore() {
-        return notBefore;
-    }
-
-    public void setNotBefore(int notBefore) {
-        this.updated |= ! Objects.equals(this.notBefore, notBefore);
-        this.notBefore = notBefore;
-    }
-
-    public Set<String> getScope() {
-        return scope;
-    }
-
-    public void setScope(Set<String> scope) {
-        this.updated |= ! Objects.equals(this.scope, scope);
-        this.scope.clear();
-        this.scope.addAll(scope);
-    }
-
-    public Set<String> getWebOrigins() {
-        return webOrigins;
-    }
-
-    public void setWebOrigins(Set<String> webOrigins) {
-        this.updated |= ! Objects.equals(this.webOrigins, webOrigins);
-        this.webOrigins.clear();
-        this.webOrigins.addAll(webOrigins);
-    }
-
-    public ProtocolMapperModel addProtocolMapper(ProtocolMapperModel model) {
-        Objects.requireNonNull(model.getId(), "protocolMapper.id");
-        updated = true;
-        this.protocolMappers.put(model.getId(), model);
-        return model;
-    }
-
-    public Collection<ProtocolMapperModel> getProtocolMappers() {
-        return protocolMappers.values();
-    }
-
-    public void updateProtocolMapper(String id, ProtocolMapperModel mapping) {
-        updated = true;
-        protocolMappers.put(id, mapping);
-    }
-
-    public void removeProtocolMapper(String id) {
-        updated |= protocolMappers.remove(id) != null;
-    }
-
-    public void setProtocolMappers(Collection<ProtocolMapperModel> protocolMappers) {
-        this.updated |= ! Objects.equals(this.protocolMappers, protocolMappers);
-        this.protocolMappers.clear();
-        this.protocolMappers.putAll(protocolMappers.stream().collect(Collectors.toMap(ProtocolMapperModel::getId, Function.identity())));
-    }
-
-    public ProtocolMapperModel getProtocolMapperById(String id) {
-        return id == null ? null : protocolMappers.get(id);
-    }
-
-    public boolean isSurrogateAuthRequired() {
-        return surrogateAuthRequired;
-    }
-
-    public void setSurrogateAuthRequired(boolean surrogateAuthRequired) {
-        this.updated |= ! Objects.equals(this.surrogateAuthRequired, surrogateAuthRequired);
-        this.surrogateAuthRequired = surrogateAuthRequired;
-    }
-
-    public String getManagementUrl() {
-        return managementUrl;
-    }
-
-    public void setManagementUrl(String managementUrl) {
-        this.updated |= ! Objects.equals(this.managementUrl, managementUrl);
-        this.managementUrl = managementUrl;
-    }
-
-    public String getRootUrl() {
-        return rootUrl;
-    }
-
-    public void setRootUrl(String rootUrl) {
-        this.updated |= ! Objects.equals(this.rootUrl, rootUrl);
-        this.rootUrl = rootUrl;
-    }
-
-    public String getBaseUrl() {
-        return baseUrl;
-    }
-
-    public void setBaseUrl(String baseUrl) {
-        this.updated |= ! Objects.equals(this.baseUrl, baseUrl);
-        this.baseUrl = baseUrl;
-    }
-
-    public boolean isBearerOnly() {
-        return bearerOnly;
-    }
-
-    public void setBearerOnly(boolean bearerOnly) {
-        this.updated |= ! Objects.equals(this.bearerOnly, bearerOnly);
-        this.bearerOnly = bearerOnly;
-    }
-
-    public boolean isConsentRequired() {
-        return consentRequired;
-    }
-
-    public void setConsentRequired(boolean consentRequired) {
-        this.updated |= ! Objects.equals(this.consentRequired, consentRequired);
-        this.consentRequired = consentRequired;
-    }
-
-    public boolean isStandardFlowEnabled() {
-        return standardFlowEnabled;
-    }
-
-    public void setStandardFlowEnabled(boolean standardFlowEnabled) {
-        this.updated |= ! Objects.equals(this.standardFlowEnabled, standardFlowEnabled);
-        this.standardFlowEnabled = standardFlowEnabled;
-    }
-
-    public boolean isImplicitFlowEnabled() {
-        return implicitFlowEnabled;
-    }
-
-    public void setImplicitFlowEnabled(boolean implicitFlowEnabled) {
-        this.updated |= ! Objects.equals(this.implicitFlowEnabled, implicitFlowEnabled);
-        this.implicitFlowEnabled = implicitFlowEnabled;
-    }
-
-    public boolean isDirectAccessGrantsEnabled() {
-        return directAccessGrantsEnabled;
-    }
-
-    public void setDirectAccessGrantsEnabled(boolean directAccessGrantsEnabled) {
-        this.updated |= ! Objects.equals(this.directAccessGrantsEnabled, directAccessGrantsEnabled);
-        this.directAccessGrantsEnabled = directAccessGrantsEnabled;
-    }
-
-    public boolean isServiceAccountsEnabled() {
-        return serviceAccountsEnabled;
-    }
-
-    public void setServiceAccountsEnabled(boolean serviceAccountsEnabled) {
-        this.updated |= ! Objects.equals(this.serviceAccountsEnabled, serviceAccountsEnabled);
-        this.serviceAccountsEnabled = serviceAccountsEnabled;
-    }
-
-    public int getNodeReRegistrationTimeout() {
-        return nodeReRegistrationTimeout;
-    }
-
-    public void setNodeReRegistrationTimeout(int nodeReRegistrationTimeout) {
-        this.updated |= ! Objects.equals(this.nodeReRegistrationTimeout, nodeReRegistrationTimeout);
-        this.nodeReRegistrationTimeout = nodeReRegistrationTimeout;
-    }
-
-    public void addWebOrigin(String webOrigin) {
-        updated = true;
-        this.webOrigins.add(webOrigin);
-    }
-
-    public void removeWebOrigin(String webOrigin) {
-        updated |= this.webOrigins.remove(webOrigin);
-    }
-
-    public void addRedirectUri(String redirectUri) {
-        this.updated |= ! this.redirectUris.contains(redirectUri);
-        this.redirectUris.add(redirectUri);
-    }
-
-    public void removeRedirectUri(String redirectUri) {
-        updated |= this.redirectUris.remove(redirectUri);
-    }
-
-    public void removeAttribute(String name) {
-        this.updated |= this.attributes.remove(name) != null;
-    }
-
-    public List<String> getAttribute(String name) {
-        return attributes.getOrDefault(name, Collections.EMPTY_LIST);
-    }
-
-    public String getAuthenticationFlowBindingOverride(String binding) {
-        return this.authFlowBindings.get(binding);
-    }
-
-    public Map<String, String> getAuthenticationFlowBindingOverrides() {
-        return this.authFlowBindings;
-    }
-
-    public void removeAuthenticationFlowBindingOverride(String binding) {
-        updated |= this.authFlowBindings.remove(binding) != null;
-    }
-
-    public void setAuthenticationFlowBindingOverride(String binding, String flowId) {
-        this.updated = true;
-        this.authFlowBindings.put(binding, flowId);
-    }
-
-    public Collection<String> getScopeMappings() {
-        return scopeMappings;
-    }
-
-    public void addScopeMapping(String id) {
-        if (id != null) {
-            updated = true;
-            scopeMappings.add(id);
-        }
-    }
-
-    public void deleteScopeMapping(String id) {
-        updated |= scopeMappings.remove(id);
-    }
-
-    public void addClientScope(String id, boolean defaultScope) {
-        if (id != null) {
-            updated = true;
-            this.clientScopes.put(id, defaultScope);
-        }
-    }
-
-    public void removeClientScope(String id) {
-        if (id != null) {
-            updated |= clientScopes.remove(id) != null;
-        }
-    }
-
-    public Stream<String> getClientScopes(boolean defaultScope) {
-        return this.clientScopes.entrySet().stream()
-          .filter(me -> Objects.equals(me.getValue(), defaultScope))
-          .map(Entry::getKey);
-    }
-
-    public String getRealmId() {
-        return this.realmId;
-    }
+public interface MapClientEntity<K> extends AbstractEntity<K> {
+
+    void addClientScope(String id, Boolean defaultScope);
+
+    ProtocolMapperModel addProtocolMapper(ProtocolMapperModel model);
+
+    void addRedirectUri(String redirectUri);
+
+    void addScopeMapping(String id);
+
+    void addWebOrigin(String webOrigin);
+
+    void deleteScopeMapping(String id);
+
+    List<String> getAttribute(String name);
+
+    Map<String, List<String>> getAttributes();
+
+    Map<String, String> getAuthFlowBindings();
+
+    String getAuthenticationFlowBindingOverride(String binding);
+
+    Map<String, String> getAuthenticationFlowBindingOverrides();
+
+    String getBaseUrl();
+
+    String getClientAuthenticatorType();
+
+    String getClientId();
+
+    Stream<String> getClientScopes(boolean defaultScope);
+
+    String getDescription();
+
+    String getManagementUrl();
+
+    String getName();
+
+    int getNodeReRegistrationTimeout();
+
+    int getNotBefore();
+
+    String getProtocol();
+
+    ProtocolMapperModel getProtocolMapperById(String id);
+
+    Collection<ProtocolMapperModel> getProtocolMappers();
+
+    String getRealmId();
+
+    Set<String> getRedirectUris();
+
+    String getRegistrationToken();
+
+    String getRootUrl();
+
+    Set<String> getScope();
+
+    Collection<String> getScopeMappings();
+
+    String getSecret();
+
+    Set<String> getWebOrigins();
+
+    Boolean isAlwaysDisplayInConsole();
+
+    Boolean isBearerOnly();
+
+    Boolean isConsentRequired();
+
+    Boolean isDirectAccessGrantsEnabled();
+
+    Boolean isEnabled();
+
+    Boolean isFrontchannelLogout();
+
+    Boolean isFullScopeAllowed();
+
+    Boolean isImplicitFlowEnabled();
+
+    Boolean isPublicClient();
+
+    Boolean isServiceAccountsEnabled();
+
+    Boolean isStandardFlowEnabled();
+
+    Boolean isSurrogateAuthRequired();
+
+    void removeAttribute(String name);
+
+    void removeAuthenticationFlowBindingOverride(String binding);
+
+    void removeClientScope(String id);
+
+    void removeProtocolMapper(String id);
+
+    void removeRedirectUri(String redirectUri);
+
+    void removeWebOrigin(String webOrigin);
+
+    void setAlwaysDisplayInConsole(Boolean alwaysDisplayInConsole);
+
+    void setAttribute(String name, List<String> values);
+
+    void setAuthFlowBindings(Map<String, String> authFlowBindings);
+
+    void setAuthenticationFlowBindingOverride(String binding, String flowId);
+
+    void setBaseUrl(String baseUrl);
+
+    void setBearerOnly(Boolean bearerOnly);
+
+    void setClientAuthenticatorType(String clientAuthenticatorType);
+
+    void setClientId(String clientId);
+
+    void setConsentRequired(Boolean consentRequired);
+
+    void setDescription(String description);
+
+    void setDirectAccessGrantsEnabled(Boolean directAccessGrantsEnabled);
+
+    void setEnabled(Boolean enabled);
+
+    void setFrontchannelLogout(Boolean frontchannelLogout);
+
+    void setFullScopeAllowed(Boolean fullScopeAllowed);
+
+    void setImplicitFlowEnabled(Boolean implicitFlowEnabled);
+
+    void setManagementUrl(String managementUrl);
+
+    void setName(String name);
+
+    void setNodeReRegistrationTimeout(int nodeReRegistrationTimeout);
+
+    void setNotBefore(int notBefore);
+
+    void setProtocol(String protocol);
+
+    void setProtocolMappers(Collection<ProtocolMapperModel> protocolMappers);
+
+    void setPublicClient(Boolean publicClient);
+
+    void setRedirectUris(Set<String> redirectUris);
+
+    void setRegistrationToken(String registrationToken);
+
+    void setRootUrl(String rootUrl);
+
+    void setScope(Set<String> scope);
+
+    void setSecret(String secret);
+
+    void setServiceAccountsEnabled(Boolean serviceAccountsEnabled);
+
+    void setStandardFlowEnabled(Boolean standardFlowEnabled);
+
+    void setSurrogateAuthRequired(Boolean surrogateAuthRequired);
+
+    void setWebOrigins(Set<String> webOrigins);
+
+    void updateProtocolMapper(String id, ProtocolMapperModel mapping);
 
 }
