+++ model/map/src/main/java/org/keycloak/models/map/client/MapClientEntityDelegate.java
@@ -1,0 +1,36 @@
+/*
+ * Copyright 2021 Red Hat, Inc. and/or its affiliates
+ * and other contributors as indicated by the @author tags.
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.keycloak.models.map.client;
+
+/**
+ *
+ * @author hmlnarik
+ */
+public class MapClientEntityDelegate<K> extends MapClientEntityLazyDelegate<K> {
+
+    private final MapClientEntity<K> delegate;
+
+    public MapClientEntityDelegate(MapClientEntity<K> delegate) {
+        super(null);
+        this.delegate = delegate;
+    }
+
+    @Override
+    protected MapClientEntity<K> getDelegate() {
+        return delegate;
+    }
+}
